Attempting to get the reaction cache working
There is a bug when attempting to copy the setup used by reagent which makes use of the subscription cache
which is a map from subscription query to a reaction.

The cache is correctly being queried but the cache is never being updated. This I don't understand,
so I'm leaving it here for now.